// modulo5/cookenu/src/types.ts
export enum USER_ROLES {
    ADMIN = "ADMIN",
    NORMAL = "NORMAL"
 }
 
 export interface authenticationData {
    id:string,
    role: USER_ROLES
 }
 
 export interface user {
    id: string
    email: string
    name: string
    password: string
    role: USER_ROLES
 }

  export interface recipe {
    title: string
    description: string
    creation_date: string
    creator_id: string
 }